9fl merge2

+++ sale_membership/models.py
@@ -25,7 +25,7 @@ class Person(models.Model):
 
     _inherit = 'res.partner'
 
-    points = fields.Integer(string='Current membership Points', default=0)
+    points = fields.Float(string='Current membership Points', default=0)
     type_id = fields.Many2one('sale_membership.type', compute='set_membership', string='Current Membership type', store=True)
     blocked = fields.Boolean(default=False, string='Blocked', readonly=True)
